REFACTOR: move read_param_fields() to args.py
